Add playwright test for suspending the service worker.
